twitter-streaming-example/requirements.txt
kafka-python==2.0.1
python-twitter
tweepy
dotenv